+++ src/http-server/src/Server.php
@@ -86,7 +86,11 @@ class Server implements OnRequestInterface, MiddlewareInitializerInterface
 
             [$psr7Request, $psr7Response] = $this->initRequestAndResponse($request, $response);
 
-            $this->option?->isEnableRequestLifecycle() && $this->event?->dispatch(new RequestReceived($psr7Request, $psr7Response));
+            $this->option?->isEnableRequestLifecycle() && $this->event?->dispatch(new RequestReceived(
+                request: $psr7Request,
+                response: $psr7Response,
+                server: $this->serverName
+            ));
 
             $psr7Request = $this->coreMiddleware->dispatch($psr7Request);
             /** @var Dispatched $dispatched */
@@ -108,9 +112,19 @@ class Server implements OnRequestInterface, MiddlewareInitializerInterface
             });
         } finally {
             if (isset($psr7Request) && $this->option?->isEnableRequestLifecycle()) {
-                defer(fn () => $this->event?->dispatch(new RequestTerminated($psr7Request, $psr7Response ?? null, $throwable ?? null)));
+                defer(fn () => $this->event?->dispatch(new RequestTerminated(
+                    request: $psr7Request,
+                    response: $psr7Response ?? null,
+                    exception: $throwable ?? null,
+                    server: $this->serverName
+                )));
 
-                $this->event?->dispatch(new RequestHandled($psr7Request, $psr7Response ?? null, $throwable ?? null));
+                $this->event?->dispatch(new RequestHandled(
+                    request: $psr7Request,
+                    response: $psr7Response ?? null,
+                    exception: $throwable ?? null,
+                    server: $this->serverName
+                ));
             }
 
             // Send the Response to client.